# backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MonkeyGraph Agent API",
    "description": "MonkeyGraph 知识图谱系统的 Agent API 接口服务文档",
    "version": "1.0.0",
    "contact": {
      "name": "MonkeyGraph Support",
      "url": "https://github.com/czhmisaka/czh_graph"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:13001",
      "description": "本地开发服务器"
    }
  ],
...
  "paths": {
...
    "/api/agent/graphs/{graphId}": {
      "get": {
        "tags": ["Graphs"],
        "summary": "获取图谱详情（精简版）",
        "description": "获取图谱基本信息，不包含 nodes 和 edges。获取图谱概览请使用 /summary 接口，获取节点请使用 /nodes 接口。",
        "security": [
          {
            "AgentAuth": []
          }
        ],
        "parameters": [
          {
            "name": "graphId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "图谱 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "graph": {
                    "id": "graph-id",
                    "name": "图谱名称",
                    "description": "描述",
                    "created_at": "2026-03-24T00:00:00.000Z"
                  }
                }
              }
            }
          },
          "404": {
            "description": "图谱不存在"
          }
        }
      },
      "delete": {
        "tags": ["Graphs"],
        "summary": "删除图谱",
        "security": [
          {
            "AgentAuth": []
          }
        ],
        "parameters": [
          {
            "name": "graphId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AgentAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "使用 'Agent your-api-key' 格式进行认证"
      }
    }
  }
}